Make history traversal more intuitive
When the user manually changes directory to a directory that was
already in history adjacent to the cursor, this is now treated as
a prev/next motion.

Example:

  @           # new shell
  @ cd BASE   # history: [~, ~/BASE]             []
  @ cd sub    # history: [~, ~/BASE, ~/BASE/sub] []
  @ cd ..     # history: [~, ~/BASE]             [~/BASE/sub]

Additionally, when the user is in the middle of history traversal
(with the cursor not pointing at the last history element), history
will be truncated when they go to an entirely new directory now.

Example:

  @           # new shell
  @ cd BASE   # history: [~, ~/BASE]             []
  @ cd sub    # history: [~, ~/BASE, ~/BASE/sub] []
  @ prevd     # history: [~, ~/BASE]             [~/BASE/sub]
  @ cd /tmp   # history: [~, ~/BASE, /tmp]       []
  @ prevd     # history: [~, ~/BASE]             [/tmp]

The previous behavior kept ~/BASE/sub in history, so invoking `prevd`
after `cd /tmp` was "returning" to ~/BASE/sub instead of the expected
~/BASE.